refactor/move RHF logic to separated hook
