+++ src/components/contact/contactForm/useContactForm.ts
@@ -1,37 +1,20 @@
-import { mutateData } from '@/utils/axios';
 import type { ContactFormValues } from '@/utils/schemas/contactFormSchema';
-import { useMutation } from 'react-query';
-
-type ContactFormPayload = Pick<ContactFormValues, 'email' | 'name' | 'message'>;
+import { contactFormSchema } from '@/utils/schemas/contactFormSchema';
+import { zodResolver } from '@hookform/resolvers/zod';
+import { useForm } from 'react-hook-form';
 
 export const useContactForm = () => {
-	const { data, isLoading, isSuccess, isError, error, mutate } = useMutation(
-		(payload: ContactFormPayload) => mutateData('/api/contact', payload)
-	);
-
-	const submit = ({
-		email,
-		name,
-		message,
-		termsOfService,
-	}: ContactFormValues) => {
-		if (!termsOfService) {
-			return;
-		}
-
-		mutate({
-			email,
-			name,
-			message,
-		});
-	};
+	const {
+		register,
+		handleSubmit,
+		formState: { errors },
+	} = useForm<ContactFormValues>({
+		resolver: zodResolver(contactFormSchema),
+	});
 
 	return {
-		submit,
-		data,
-		isLoading,
-		isSuccess,
-		isError,
-		error,
+		register,
+		handleSubmit,
+		errors,
 	};
 };
